Move colors and icons/ to common/
