+++ components/common/colors.ts
