+++ components/common/icons/double-palm-tree.jsx
