+++ components/common/tooltips/PlfTooltip.jsx
@@ -1,4 +1,4 @@
-import { plfColor } from "../../colors";
+import { plfColor } from "../colors";
 import { createColoredTooltip } from "./create-colored-tooltip";
 
 export const PlfTooltip = createColoredTooltip(plfColor);
