+++ components/common/tooltips/ReformTooltip.jsx
@@ -1,4 +1,4 @@
-import { reformColor } from "../../colors";
+import { reformColor } from "../colors";
 import { createColoredTooltip } from "./create-colored-tooltip";
 
 export const ReformTooltip = createColoredTooltip(reformColor, "#000000");
