+++ components/ir/cartes-impact/simple-card/simple-card-component.jsx
@@ -17,7 +17,7 @@ import PropTypes from "prop-types";
 import React, { Fragment } from "react";
 
 import { Card } from "../../../common";
-import DoublePalmTreeIcon from "../../../icons/double-palm-tree";
+import DoublePalmTreeIcon from "../../../common/icons/double-palm-tree";
 import { NeutralTooltip } from "../../../common";
 import { formatNumber } from "../../../common";
 import SimpleCardImpactImpots from "./impact-impots";
